+++ test/serve/basic/function-with-argv.config.js
@@ -1,4 +1,5 @@
 const WebpackCLITestPlugin = require('../../utils/webpack-cli-test-plugin');
+const { isDevServer4 } = require('../../utils/test-utils');
 
 module.exports = (env, argv) => {
     console.log(argv);
@@ -7,5 +8,12 @@ module.exports = (env, argv) => {
         mode: 'development',
         devtool: false,
         plugins: [new WebpackCLITestPlugin(['mode'], false, 'hooks.compilation.taps')],
+        devServer: isDevServer4
+            ? {
+                  client: {
+                      logging: 'info',
+                  },
+              }
+            : {},
     };
 };